+++ src/main/java/domain/card/Card.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.card;
+
+import domain.Score;
 
 public class Card {
