+++ src/main/java/domain/card/Hand.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.card;
+
+import domain.Score;
 
 import java.util.List;
