+++ src/test/java/domain/DeckTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.card.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
